// libs/features/establishment/src/lib/shared/enums/establishment-status-error-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum EstablishmentStatusErrorEnum {
  OPENING_IN_PROGRESS = 'ESTABLISHMENT.ERROR.OPENING-IN-PROGRESS',
  UNDER_INSPECTION = 'ESTABLISHMENT.ERROR.OPENING-IN-PROGRESS',
  CLOSING_IN_PROGRESS = 'ESTABLISHMENT.ERROR.CLOSING-IN-PROGRESS',
  CLOSED = 'ESTABLISHMENT.ERROR.CLOSED',
  CANCELLED = 'ESTABLISHMENT.ERROR.CANCELLED',
  GROUP_ACCESS_CANCELLED = 'ESTABLISHMENT.ERROR.GROUP-ACCESS-CANCELLED',
  GROUP_ACCESS_DRAFT = 'ESTABLISHMENT.ERROR.GROUP-ACCESS-DRAFT',
  CANCEL_UNDER_INSPECTION = 'ESTABLISHMENT.ERROR.CANCELLED',
  REOPENING_IN_PROGRESS = 'ESTABLISHMENT.ERROR.REOPENING-IN-PROGRESS',
  UNDER_CLOSURE_WAITING_SETTLEMENT = 'ESTABLISHMENT.ERROR.UNDER-CLOSURE-WAITING-SETTLEMENT',
  OPENING_IN_PROGRESS_INT = 'ESTABLISHMENT.ERROR.OPENING-IN-PROGRESS-INT',
  DRAFT = 'ESTABLISHMENT.WARNING.NOT-REGISTERED',
  PENDING = 'ESTABLISHMENT.ERROR.PENDING',
  OPENING_IN_PROGRESS_GOL_UPDATE = 'ESTABLISHMENT.ERROR.OPENING-IN-PROGRESS-GOL-UPDATE'
}