// libs/features/establishment/src/lib/shared/enums/establishment-trans-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum EstablishmentTransEnum {
  LINK_TRANSACTION = 300319,
  DELINK_TRANSACTION = 300318,
  BASIC_DETAILS_TRANSACTION = 300311,
  OWNER_UPDATE_TRANSACTION = 300313,
  LE_CHANGE = 300314,
  ZAKAT_CERTIFICATE = 300357,
  GOOD_PAYMENT_CERTIFCATE = 300356,
  SAFETY_INSPECTION_TRANSACTION = 300344,
  IDENTIER_UPDATE_TRANSACTION = 300310,
  ADDRESS_UPDATE_TRANSACTION = 300315,
  BANK_DETAILS_UPDATE = 300312,
  CONTACT_DETAILS_UPDATE = 300316,
  LATE_FEE_UPDATE = 300329,
  PAYMENT_TYPE_UPDATE = 300353,
  CLOSE_ESTABLISHMENT_TRANSACTION = 300321,
  CHANGE_BRANCH_TO_MAIN = 300317,
  ADD_FLAG_TRANSACTION = 300322,
  MODIFY_FLAG_TRANSACTION = 300324,
  RASED_DOCUMENT_UPLOAD = 300350,
  SUPER_ADMIN_ADD_TRANSACTION = 300337,
  GCC_ADMIN_ADD_TRANSACTION = 300380,
  OTHER_ADMINS_ADD_TRANSACTION = 300327,
  DELETE_ADMINS_TRANSACTION = 300354,
  MODIFY_ADMINS_TRANSACTION = 300328,
  REPLACE_SUPER_ADMIN_TRANSACTION = 300320,
  REPLACE_GCC_ADMIN_TRANSACTION = 300359,
  REPLACE_BRANCH_ROLE_ADMIN_TRANSACTION = 300336,
  REGISTER_ESTABLISHMENT_TRANSACTION = 100613,
  ADD_RELATIONSHIP_MANAGER = 300396,
  MODIFY_RELATIONSHIP_MANAGER = 300395,
  REOPON_ESTABLISHMENT_TRANSACTION = 300398,
  SC_EVALUATION_TRANSACTION = 300408,
  OH_CERTIFICATE = 300384,
  COMPLETE_PROACTIVE_TRANSACTION = 300338,
  MEDICAL_INSURANCE_ESTABLISHMENT_ENROLLMENT = 300402
}