// libs/features/establishment/src/lib/shared/enums/legal-entity-enum.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export enum LegalEntityEnum {
  GOVERNMENT = 'Government',
  SEMI_GOV = 'Semi Government',
  ORG_REGIONAL = 'Organization regional or international',
  INDIVIDUAL = 'Individual',
  SOCIETY = 'Society',
  PARTNERSHIP = 'Partnership',
  LIMITED_LIABILITY = 'Limited Liability',
  SHARE_STOCK_COMPANY = 'Stock Share Company',
  LIMITED_PARTNERSHIP = 'Limited Partnership',
  VOCATIONAL_ESTABLISHMENT = 'Vocational_Establishment'
}