// libs/features/establishment/src/lib/shared/models/person-search-query-param.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */

export class PersonSearchQueryParam {
  role: string = undefined;
  birthDate: string = undefined;
  personType: string = undefined;
  NIN: number = undefined;
  nationality: string = undefined;
  gccId: number = undefined;
  iqamaNo: number = undefined;
  passportNo: string = undefined;
}